python_agents/requirements.txt
langgraph==0.2.55
langchain==0.3.13
langchain-google-genai==2.0.5
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3